+++ NEWS.md
@@ -1,4 +1,4 @@
-# insight (devel)
+# insight 1.5.4
 
 ## New supported models
 
@@ -19,6 +19,8 @@
   (df) instead of the t value for `rlmerMod` models (*robustlmm*) after
   Satterthwaite degrees of freedom were cached (e.g., following `emmeans`
   calls).
+
+* Fix failing CRAN checks.
 
 # insight 1.5.3